requirements.txt
python-dotenv==1.0.0
python-telegram-bot==22.1.0
